Move harvest logic to separate file
